# openapi-3.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jiro Search API",
    "version": "0.2.0",
    "description": "Local-first web search, scraping, and social media intelligence platform. Supports 9+ search engines, 12 social platforms, hybrid search, structured extraction, and smart intent routing.",
    "contact": {
      "name": "Jiro Team",
      "url": "https://github.com/DevAnimecx/jiro"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.jiro.dev",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/smart": {
      "post": {
        "tags": [
          "smart"
        ],
        "summary": "Smart search with intent routing",
        "description": "Automatically detect search intent and route to the appropriate handler.",
        "operationId": "smartSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Smart search result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/smart/classify": {
      "post": {
        "tags": [
          "smart"
        ],
        "summary": "Classify search intent",
        "description": "Classify the intent of a search query without executing it.",
        "operationId": "smartClassify",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartClassifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intent classification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartClassifyResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}